requirements.txt
Django>=4.2,<5.0
psycopg2-binary
python-dotenv
stripe>=12.0.0,<13.0.0
google-cloud-tasks>=2.13.0,<3.0.0
whitenoise
google-cloud-storage==2.16.0